backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
aiosmtplib
pydantic
email-validator